Try again.

+++ source/MRDotNet2Test/src/ExpandShrink.cs
@@ -13,7 +13,7 @@ namespace MRTest
             var region = expand(mesh.topology, new FaceId(0), 3);
             Assert.That(region.count, Is.EqualTo((nuint)75));
             expand(mesh.topology, region, 3);
-            Assert.That(region.count, Is.GreaterThan(75));
+            Assert.That(region.count, Is.GreaterThan((nuint)75));
             shrink(mesh.topology, region, 3);
             Assert.That(region.count, Is.EqualTo((nuint)75));
         }
@@ -25,7 +25,7 @@ namespace MRTest
             var region = expand(mesh.topology, new VertId(0), 3);
             Assert.That(region.count, Is.EqualTo((nuint)37));
             expand(mesh.topology, region, 3);
-            Assert.That(region.count, Is.GreaterThan(37));
+            Assert.That(region.count, Is.GreaterThan((nuint)37));
             shrink(mesh.topology, region, 3);
             Assert.That(region.count, Is.EqualTo((nuint)37));
         }
